requirements.txt
torch~=1.13.0
torchvision~=0.14.0
torchaudio~=0.13.1
ddu_dirty_mnist~=1.1.1
tqdm~=4.64.1
numpy~=1.23.4
matplotlib~=3.6.2
Pillow~=9.3.0
requests~=2.28.1
PyInquirer~=1.0.3